+++ TestVillage.py
@@ -1,25 +1,25 @@
 import unittest
 import unittest.mock
-from Tile import Tile
+from Village import Village
 
 
-class TestTile(unittest.TestCase):
+class TestVillage(unittest.TestCase):
 
     most = 6
     initial = 0
     lowest = 0
 
     def setUp(self):
-        self.test = Tile(TestTile.most)
+        self.test = Village(TestVillage.most)
 
     def tearDown(self):
         del self.test
 
     def test_pillagers_init(self):
-        self.assertEqual(self.test.pillagers, TestTile.initial)
+        self.assertEqual(self.test.pillagers, TestVillage.initial)
 
     def test_defenders_init(self):
-        self.assertEqual(self.test.defenders, TestTile.initial)
+        self.assertEqual(self.test.defenders, TestVillage.initial)
 
     def test_desolated_init(self):
         self.assertFalse(self.test.desolated)
@@ -30,17 +30,22 @@ class TestTile(unittest.TestCase):
     def test_start_init(self):
         self.assertFalse(self.test.start)
 
-    def test_max_init(self):
-        self.assertEqual(self.test.most, TestTile.most)
+    def test_most_init(self):
+        self.assertEqual(self.test.most, TestVillage.most)
+
+    def test_trade_init(self):
+        self.assertFalse(self.test.trade)
 
     def test_str(self):
         self.assertEqual(str(self.test), (f"""
-                pillage_value = {Tile.pillage_value},
+                pillage_value = {self.test.pillage_value},
                 defenders = {self.test.defenders},
                 pillagers = {self.test.pillagers},
                 desolated = {self.test.desolated},
                 abandoned = {self.test.abandoned},
-                start tile = {self.test.start}"""))
+                start tile = {self.test.start},
+                trade = {self.test.trade},
+                trade_value = {self.test.trade_value}"""))
 
     def test_change_pillagers_most(self):
         result = self.test.change_pillagers(self.most-self.initial)
@@ -109,21 +114,21 @@ class TestTile(unittest.TestCase):
         self.test.desolated = True
         self.assertEqual(self.test.pillage(), {})
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
     def test_pillage_abandoned_false(self, mock_roll):
         self.test.pillagers = 1
         self.test.abandoned = True
         self.assertEqual(self.test.pillage(), {})
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=most//2+1)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=most//2+1)
     def test_pillage_abandoned_return(self, mock_roll):
         self.test.pillagers = 1
         self.test.abandoned = True
-        self.assertEqual(self.test.pillage(), Tile.pillage_value)
+        self.assertEqual(self.test.pillage(), Village.pillage_value)
 
     def test_pillage_pass(self):
         self.test.pillagers = 1
-        self.assertEqual(self.test.pillage(), Tile.pillage_value)
+        self.assertEqual(self.test.pillage(), Village.pillage_value)
 
     def test_roll_die(self):
         count = 100
@@ -135,26 +140,26 @@ class TestTile(unittest.TestCase):
             count -= 1
         self.assertTrue(output)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
     def test_check_defence_low(self, mock_roll):
         self.test.check_defence()
         self.assertEqual(self.test.defenders, self.initial)
         self.assertEqual(self.test.pillagers, self.initial)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=most)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=most)
     def test_check_defence_most_die(self, mock_roll):
         self.test.check_defence()
         self.assertEqual(self.test.defenders, self.initial)
         self.assertEqual(self.test.pillagers, self.initial)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
     def test_check_defence_low_most_defenders(self, mock_roll):
         self.test.defenders = self.most-1
         self.test.check_defence()
         self.assertEqual(self.test.defenders, self.most-1)
         self.assertEqual(self.test.pillagers, self.initial)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
     def test_check_defence_over(self, mock_roll):
         self.test.pillagers = 1
         self.test.defenders = self.most
@@ -162,7 +167,7 @@ class TestTile(unittest.TestCase):
         self.assertEqual(self.test.defenders, self.most-1)
         self.assertEqual(self.test.pillagers, 0)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
     def test_check_defence_over_extra_pillagers(self, mock_roll):
         self.test.pillagers = 2
         self.test.defenders = self.most
@@ -170,7 +175,7 @@ class TestTile(unittest.TestCase):
         self.assertEqual(self.test.defenders, self.most-1)
         self.assertEqual(self.test.pillagers, 2)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=2)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=2)
     def test_check_defence_over_two(self, mock_roll):
         self.test.pillagers = 1
         self.test.defenders = self.most
@@ -178,7 +183,7 @@ class TestTile(unittest.TestCase):
         self.assertEqual(self.test.defenders, self.most-2)
         self.assertEqual(self.test.pillagers, 0)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=3)
+    @unittest.mock.patch.object(Village, "roll_die", return_value=3)
     def test_check_defence_complex(self, mock_roll):
         self.test.pillagers = 3
         self.test.defenders = self.most
@@ -186,38 +191,71 @@ class TestTile(unittest.TestCase):
         self.assertEqual(self.test.defenders, self.most-3)
         self.assertEqual(self.test.pillagers, 2)
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
-    @unittest.mock.patch.object(Tile, "abandon")
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "abandon")
     def test_check_abandon_pass(self, mock_abandon, mock_roll):
         self.test.check_abandon(0)
         mock_abandon.assert_not_called()
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=most)
-    @unittest.mock.patch.object(Tile, "abandon")
+    @unittest.mock.patch.object(Village, "roll_die", return_value=most)
+    @unittest.mock.patch.object(Village, "abandon")
     def test_check_abandon_with_fear(self, mock_abandon, mock_roll):
         fear = self.most - 1
         self.test.check_abandon(fear)
         mock_abandon.assert_not_called()
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=1)
-    @unittest.mock.patch.object(Tile, "abandon")
+    @unittest.mock.patch.object(Village, "roll_die", return_value=1)
+    @unittest.mock.patch.object(Village, "abandon")
     def test_check_abandon_least(self, mock_roll, mock_abandon):
         self.test.check_abandon(1)
         mock_abandon.assert_called()
 
-    @unittest.mock.patch.object(Tile, "roll_die", return_value=most)
-    @unittest.mock.patch.object(Tile, "abandon")
+    @unittest.mock.patch.object(Village, "roll_die", return_value=most)
+    @unittest.mock.patch.object(Village, "abandon")
     def test_check_abandon_most(self, mock_roll, mock_abandon):
         self.test.check_abandon(self.most)
         mock_abandon.assert_called()
 
-    @unittest.mock.patch.object(Tile, "check_defence")
-    @unittest.mock.patch.object(Tile, "pillage", return_value={})
-    @unittest.mock.patch.object(Tile, "check_abandon")
-    @unittest.mock.patch.object(Tile, "change_pillagers")
-    def test_complete_turn(self, mock_ps, mock_a, mock_p, mock_d):
+    @unittest.mock.patch.object(Village, "check_defence")
+    @unittest.mock.patch.object(Village, "pillage", return_value={})
+    @unittest.mock.patch.object(Village, "check_abandon")
+    @unittest.mock.patch.object(Village, "change_pillagers")
+    @unittest.mock.patch.object(Village, "check_trade", return_value={})
+    def test_complete_turn_trade_false(self, mock_t, mock_ps, mock_a, mock_p, mock_d):  # noqa: E501
         self.assertEqual(self.test.complete_turn(0), {"pillagers": 0})
         mock_ps.assert_called()
         mock_a.assert_called()
         mock_p.assert_called()
         mock_d.assert_called()
+        mock_t.assert_not_called()
+
+    @unittest.mock.patch.object(Village, "check_defence")
+    @unittest.mock.patch.object(Village, "pillage", return_value={})
+    @unittest.mock.patch.object(Village, "check_abandon")
+    @unittest.mock.patch.object(Village, "change_pillagers")
+    @unittest.mock.patch.object(Village, "check_trade", return_value={})
+    def test_complete_turn_trade_true(self, mock_t, mock_ps, mock_a, mock_p, mock_d):  # noqa: E501
+        self.test.trade = True
+        self.assertEqual(self.test.complete_turn(0), {"pillagers": 0})
+        mock_ps.assert_called()
+        mock_a.assert_called()
+        mock_p.assert_not_called()
+        mock_d.assert_called()
+        mock_t.assert_called()
+
+    def test_check_trade_no_pillagers(self):
+        self.assertEqual(self.test.check_trade(), {})
+
+    def test_check_trade_desolated(self):
+        self.test.pillagers = 1
+        self.test.desolated = True
+        self.assertEqual(self.test.check_trade(), {})
+
+    def test_check_trade_abandoned_false(self):
+        self.test.pillagers = 1
+        self.test.abandoned = True
+        self.assertEqual(self.test.check_trade(), {})
+
+    def test_check_trade_pass(self):
+        self.test.pillagers = 1
+        self.assertEqual(self.test.check_trade(), Village.trade_value)